refactor: moved the get_raw_block method to the helpers mod
